Move data file. Remove old graph and data.
