+++ src/components/new-simulation/simulation-graph.tsx
@@ -4,7 +4,7 @@ import React, { ReactElement } from "react";
 import { clsx } from "clsx";
 import { useAppContext } from "../../hooks/use-app-context";
 import { useCommonState } from "../../hooks/use-common-state";
-import { getData, getSVGPath } from "./simulation-data";
+import { getData, getSVGPath } from "../../data/graph-data";
 
 const 
   dataWidth = 275,
