+++ src/components/new-simulation/simulation-results.scss
@@ -39,7 +39,7 @@
     border-radius: 8px;
 
     label {
-      color: vars.$sim-dark-1;
+      color: #04708b;
       font-weight: bold;
     }
