--- src/components/sim-graph.scss
@@ -1,47 +1,0 @@
-@use "./vars.scss";
-
-.sim-graph-container {
-  display: flex;
-  height: 141px;
-  background-color: vars.$sim-light-8;
-  font-size: 11px;
-
-  .vertical-label {
-    width: 39px;
-    height: 125px;
-    writing-mode: tb-rl;
-    transform: rotate(-180deg);
-    font-weight: bold;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .sim-graph-right {
-    display: flex;
-    flex-direction: column;
-
-    padding-top: 15px;
-    padding-right: 20px;
-
-    .sim-graph {
-      width: 307px;
-      height: 95px;
-    }
-
-    .horizontal-label {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      width: 307px;
-      height: 31px;
-
-      .bold {
-        font-family: vars.$font-bold;
-        font-weight: bold;
-      }
-    }
-  }
-}
